chore: rework helpers files and exports
