+++ lib/helpers/bucket-auto/group-by-auto-payload.ts
