+++ lib/helpers/commons/field.test.ts
@@ -1,0 +1,7 @@
+import {Field} from "./field";
+
+describe('Field', () => {
+    it('should create a javascript object', () => {
+        expect(Field('unit', 'test')).toEqual({ unit: 'test' });
+    });
+});
