+++ lib/helpers/commons/list.test.ts
@@ -1,0 +1,7 @@
+import {List} from "./list";
+
+describe('List', () => {
+    it('should create a list with the arguments passed in parameter', () => {
+        expect(List('unit', 'test')).toEqual(['unit', 'test']);
+    });
+});
