+++ lib/helpers/commons/list.ts
@@ -1,0 +1,6 @@
+/**
+ * Create a list with the elements passed in parameter
+ * @param args
+ * @constructor
+ */
+export const List = (...args: any[]) => args;
