+++ lib/helpers/current-op/op-payload.ts
