+++ lib/helpers/index.ts
@@ -1,16 +1,21 @@
 /*
 COMMON HELPERS
 */
-export {List, Field} from './commons';
+export {Field} from './commons/field';
+export {List} from './commons/list';
 
 /*
 STAGE HELPERS
 */
-export {GroupByPayload} from "./bucket";
-export {GroupByAutoPayload} from "./bucket-auto";
-export {ConditionPayload, EqualityPayload} from "./lookup";
-export {IntoPayload} from "./merge";
-export {DbCollPayload} from "./out";
-export {IgnorePayload, Ignore, OnlyPayload, Only} from "./project";
-export {SizePayload} from "./sample";
-export {CollectionPayload} from "./union-with";
+export {GroupByPayload} from "./bucket/group-by-payload";
+export {GroupByAutoPayload} from "./bucket-auto/group-by-auto-payload";
+export {ConditionPayload} from "./lookup/condition-payload";
+export {EqualityPayload} from "./lookup/equality-payload";
+export {IntoPayload} from "./merge/into-payload";
+export {DbCollPayload} from "./out/db-coll-payload";
+export {IgnorePayload} from "./project/ignore-payload";
+export {OnlyPayload} from "./project/only-payload";
+export {Ignore} from "./project/ignore";
+export {Only} from "./project/only";
+export {SizePayload} from "./sample/size-payload";
+export {CollectionPayload} from "./union-with/collection-payload";
