+++ lib/helpers/lookup/condition-payload.ts
@@ -1,30 +1,7 @@
-import { PipelineError } from "../../errors";
 import {
-    LookupConditionInterface, LookupEqualityInterface, StageInterface
+    LookupConditionInterface, StageInterface
 } from "../../interfaces";
 
-/**
- * Equality Match
- * To perform an equality match between a field from the input documents with a field from the documents of the “joined”
- * collection
- *
- * @param from
- * @param as
- * @param localField
- * @param foreignField
- */
-export const EqualityPayload = (from: string, as: string, localField: string, foreignField: string) => {
-    if (!localField || !foreignField) {
-        throw new PipelineError('Invalid Lookup Equality Payload, invalid or missing "localField" or "foreignField" parameter!');
-    }
-
-    return {
-        from: from,
-        localField: localField,
-        foreignField: foreignField,
-        as: as
-    } as LookupEqualityInterface;
-};
 /**
  * Join Conditions and Uncorrelated Sub-queries
  * To perform uncorrelated subQueries between two collections as well as allow other join conditions besides a single
