+++ lib/helpers/merge/into-payload.ts
